Add PowerAggregation
