+++ mlbench_core/utils/pytorch/distributed.py
@@ -1,5 +1,7 @@
 import torch
 import torch.distributed as dist
+import numpy as np
+from mlbench_core.utils.pytorch.utils import orthogonalize
 
 # TODO: those 3 funtions are never used, maybe delete them ?
 
@@ -272,6 +274,163 @@ class SparsifiedAggregation(Aggregation):
 
     def _agg(self, data, op):
         pass
+
+
+class PowerAggregation(Aggregation):
+    """Aggregate updates using power iteration and error feedback.
+
+    Args:
+            model (:obj:`nn.Module`): Model which contains parameters for SGD
+            use_cuda (bool): Whether to use cuda tensors for aggregation
+            reuse_query (bool): Whether to use warm start to initialize the power iteration
+            rank (int): The rank of the gradient approximation
+    """
+
+    def __init__(self, model, use_cuda=False, reuse_query=False, rank=1):
+        super(PowerAggregation, self).__init__(use_cuda=use_cuda)
+        self.p_memory = None
+        self.q_memory = None
+        self.reuse_query = reuse_query
+        self.device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+        self.rng = np.random.RandomState(1)
+        self.n_workers = dist.get_world_size()
+        self.rank = rank
+        self.memories = [torch.zeros_like(param) for param in model.parameters()]
+        self.send_buffers = [torch.zeros_like(param) for param in model.parameters()]
+
+    def set_random(self, vector):
+        """Sets the data in `vector` to random values."""
+        torch.manual_seed(self.rng.randint(1_000_000_000))
+        vector.data[:] = torch.randn(*vector.shape, device=self.device)
+
+    def _agg_gradients_by_model(self, model, op):
+        """Aggregate models gradients, all layers at once
+
+        Args:
+            model (:obj:`torch.Module`): Models to be averaged.
+            op (str): Aggregation methods like `avg`, `sum`, `min`, `max`, etc.
+        """
+        grads = [t.grad.data for t in model.parameters()]
+        aggregated = self._agg(grads, op=op)
+
+        for i, param in enumerate(model.parameters()):
+            param.grad.data = aggregated[i]
+
+    def _agg(self, data, op):
+        """Aggregate data using `op` operation.
+
+        Args:
+            data (:obj:`torch.Tensor`): A Tensor to be aggragated.
+            op (str): Aggregation methods like `avg`, `sum`, `min`, `max`, etc.
+
+        Returns:
+            :obj:`torch.Tensor`: An aggregated tensor.
+        """
+        if op == "avg":
+            for grad, memory, send_bfr in zip(data, self.memories, self.send_buffers):
+                send_bfr.data[:] = grad + memory
+            self.reduce(self.send_buffers, data, self.memories)
+        else:
+            raise NotImplementedError("op {} is not supported yet.".format(op))
+        return data
+
+    def reduce(self, grad_in, grad_out, memory_out):
+        """Reduces the gradients between the workers in place and calculates error feedback.
+
+        Args:
+            grad_in (list[torch.Tensor]): The gradients to reduce.
+            grad_out (list[torch.Tensor]): Used for storing the reduced gradients.
+            memory_out (list[torch.Tensor]): Used for storing error feedback.
+        """
+        # Split the tensors into rank1-ones that will be reduced un-compressed
+        # and rank > 1 tensors that are compressed
+        rank1_tensors = [
+            (tensor, out, mem)
+            for tensor, out, mem in zip(grad_in, grad_out, memory_out)
+            if tensor.ndimension() <= 1
+        ]
+        high_rank_tensors = [
+            (tensor, out, mem)
+            for tensor, out, mem in zip(grad_in, grad_out, memory_out)
+            if tensor.ndimension() > 1
+        ]
+
+        # We are building a rank-1 approximation of every tensor
+        # that can be interpreted as a matrix. Let the approximation be
+        # M = p q^T
+        # We are allocating consequtive memory for the p's and q's
+
+        memory_is_uninitialized = self.p_memory is None
+
+        p_total_size = 0
+        q_total_size = 0
+        for tensor, _, _ in high_rank_tensors:
+            matrix = tensor.view(tensor.shape[0], -1)
+            n, m = matrix.shape
+            rank = min(n, m, self.rank)
+            p_total_size += n * rank
+            q_total_size += m * rank
+        if self.p_memory is None:
+            self.p_memory = torch.empty(p_total_size, device=self.device)
+            self.q_memory = torch.empty(q_total_size, device=self.device)
+
+        # Find them again and make lists of pointers
+        ps = []
+        qs = []
+        p_idx = 0
+        q_idx = 0
+        for tensor, _, _ in high_rank_tensors:
+            matrix = tensor.view(tensor.shape[0], -1)
+            n, m = matrix.shape
+            rank = min(n, m, self.rank)
+            ps.append(self.p_memory[p_idx : p_idx + n * rank].view(n, rank))
+            qs.append(self.q_memory[q_idx : q_idx + m * rank].view(m, rank))
+            p_idx += n * rank
+            q_idx += m * rank
+
+        for (tensor, _, _), q, p in zip(high_rank_tensors, qs, ps):
+            matrix = tensor.view(tensor.shape[0], -1)
+            n, m = matrix.shape
+
+            if self.reuse_query and not memory_is_uninitialized:
+                pass
+            else:
+                # Sample a query vector q
+                self.set_random(q)
+
+        for (tensor, _, _), q, p in zip(high_rank_tensors, qs, ps):
+            matrix = tensor.view(tensor.shape[0], -1)
+            torch.matmul(matrix, q, out=p)
+
+        dist.all_reduce(self.p_memory)
+
+        # Start communicating rank 1 tensors
+        rank1_packed, rank1_indices, rank1_sizes = pack_tensors(
+            [tensor for (tensor, _, _) in rank1_tensors]
+        )
+
+        rank1_handle = dist.all_reduce(rank1_packed, async_op=True)
+
+        for p in ps:
+            orthogonalize(p)
+
+        for p, q, (tensor, _, _) in zip(ps, qs, high_rank_tensors):
+            matrix = tensor.view(tensor.shape[0], -1)
+            torch.matmul(matrix.t(), p, out=q)
+
+        dist.all_reduce(self.q_memory)
+        self.q_memory.data[:] /= self.n_workers
+
+        for p, q, (tensor, out, mem) in zip(ps, qs, high_rank_tensors):
+            # Set the output gradient
+            torch.matmul(p, q.t(), out=out.data[:])
+            mem.data[:] = tensor - out
+
+        rank1_handle.wait()
+        rank1_packed /= self.n_workers
+        rank1_unpacked = unpack_tensors(rank1_packed, rank1_indices, rank1_sizes)
+        for i, (_, out, _) in enumerate(rank1_tensors):
+            out[:] = rank1_unpacked[i]
 
 
 def get_backend_tensor(tensor):
